PART 4 REST 방식과 Ajax를 이용한 댓글 처리 TDD 398p study

+++ src/main/java/org/zerock/mapper/ReplyMapper.java
@@ -8,11 +8,11 @@ import org.zerock.domain.ReplyVO;
 
 public interface ReplyMapper {
 
-	public Long insert(ReplyVO vo);
+	public int insert(ReplyVO vo);
 	
 	public ReplyVO read(Long bno);
 	
-	public int delete (int rno);
+	public int delete (Long rno);
 	
 	public int update(ReplyVO reply);
